pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "RAG chatbot using LangChain + LangSmith + Ollama"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "langsmith>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pypdf>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_assistant"]